First stab at L2 bridging app
Tried with IPv4 hosts and host discovery doesn't work properly as only
ARP requests are intercepted, but not replies.

Since we support cloning to the CPU only when a packet is processed by a
multicast group, ARP replies will not be cloned as the eth dst is not
the broadcast one. It seems it's time to add support for clone sessions
in ONOS.

For now, for ping to work, both hosts need to send an ARP request with
eth dst the broadcast address.